claude-const-maximizer/backend/templates/boilerplates/backend/ai-powered-blockchain-smart-contract-analyzer/app/main.py
from fastapi import FastAPI, HTTPException, Depends
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import List, Optional, Dict, Any
import os
from dotenv import load_dotenv

# Import project-specific modules
from .database import get_db
from .models import Base, User, Contract, Analysis, Vulnerability
from .schemas import ContractCreate, AnalysisCreate, VulnerabilityCreate
from .auth import get_current_user
from .config import settings

load_dotenv()

app = FastAPI(
    title="AI-Powered Blockchain Smart Contract Analyzer",
    description="Smart contract analysis system with AI-powered security and vulnerability detection",
    version="1.0.0"
)

# CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=settings.ALLOWED_ORIGINS,
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Project-specific Pydantic models
class LoginRequest(BaseModel):
    username: str
    password: str

class LoginResponse(BaseModel):
    access_token: str
    token_type: str = "bearer"


class ContractAnalysisRequest(BaseModel):
    contract_code: str
    blockchain_type: str
    contract_language: str
    analysis_depth: str
    security_checks: List[str]

class ContractAnalysisResult(BaseModel):
    analysis_id: int
    contract_id: int
    security_score: float
    vulnerabilities_found: List[Dict[str, Any]]
    risk_level: str
    recommendations: List[str]
    ai_confidence: float

# Root endpoint
@app.get("/")
async def root():
    return {"message": "AI-Powered Blockchain Smart Contract Analyzer API"}

# Authentication endpoint
@app.post("/auth/login", response_model=LoginResponse)
async def login(request: LoginRequest):
    """Authenticate user"""
    # TODO: Implement authentication logic
    pass

# Analyze smart contract with AI-powered security analysis
@app.post("/analysis/analyze-contract", response_model=ContractAnalysisResult)
async def _analysis_analyze_contract(
    current_user = Depends(get_current_user)
):
    """Analyze smart contract with AI-powered security analysis"""
    # TODO: Implement analyze smart contract with ai-powered security analysis
    pass

# Create a new smart contract for analysis
@app.post("/contracts/", response_model=ContractAnalysisResult)
async def _contracts_(
    current_user = Depends(get_current_user)
):
    """Create a new smart contract for analysis"""
    # TODO: Implement create a new smart contract for analysis
    pass

# Get contract vulnerabilities
@app.get("/contracts/{contract_id}/vulnerabilities", response_model=ContractAnalysisResult)
async def _contracts_contract_id_vulnerabilities(
    current_user = Depends(get_current_user)
):
    """Get contract vulnerabilities"""
    # TODO: Implement get contract vulnerabilities
    pass

# Generate security analysis report
@app.get("/analysis/{analysis_id}/report", response_model=ContractAnalysisResult)
async def _analysis_analysis_id_report(
    current_user = Depends(get_current_user)
):
    """Generate security analysis report"""
    # TODO: Implement generate security analysis report
    pass

# Get common vulnerability patterns
@app.get("/vulnerabilities/patterns", response_model=ContractAnalysisResult)
async def _vulnerabilities_patterns(
    current_user = Depends(get_current_user)
):
    """Get common vulnerability patterns"""
    # TODO: Implement get common vulnerability patterns
    pass

# Get contract audit trail
@app.get("/contracts/{contract_id}/audit-trail", response_model=ContractAnalysisResult)
async def _contracts_contract_id_audit_trail(
    current_user = Depends(get_current_user)
):
    """Get contract audit trail"""
    # TODO: Implement get contract audit trail
    pass

# Health check
@app.get("/health")
async def health_check():
    return {"status": "healthy", "service": "ai-powered-blockchain-smart-contract-analyzer"}

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)


