claude-const-maximizer/backend/templates/boilerplates/backend/ai-powered-blockchain-smart-contract-analyzer/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1

# Background tasks and queues
celery==5.3.4
flower==2.0.1

# HTTP client and testing
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1

# AI and NLP libraries
openai==1.3.7
anthropic==0.7.8
spacy==3.7.2
nltk==3.8.1
transformers==4.36.0
torch==2.1.1
sentence-transformers==2.2.2

# Blockchain and Web3 integration
web3==6.11.3
eth-account==0.9.0
eth-utils==2.3.0
eth-typing==3.5.0
eth-hash==0.5.2

# Smart contract analysis and security
smart-contract-analyzer==0.1.0
security-auditor==0.1.0
vulnerability-scanner==0.1.0
code-analyzer==0.1.0

# Solidity and smart contract tools
py-solc-x==1.1.1
eth-brownie==1.19.3
hardhat==0.1.0
truffle==0.1.0

# DeFi and token analysis
defi-analyzer==0.1.0
token-analyzer==0.1.0
liquidity-analyzer==0.1.0
yield-analyzer==0.1.0

# Blockchain data and APIs
etherscan-api==0.1.0
infura-api==0.1.0
alchemy-api==0.1.0
blockchain-api==0.1.0

# Cryptography and security
cryptography==41.0.7
pycryptodome==3.19.0
hashlib==0.1.0
keccak==0.1.0

# Data processing and analysis
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
scipy==1.11.4

# Machine learning for blockchain
blockchain-ml==0.1.0
anomaly-detection==0.1.0
fraud-detection==0.1.0
risk-assessment==0.1.0

# Time series analysis for blockchain data
prophet==1.1.4
pmdarima==2.0.4
tslearn==0.6.2
arch==6.2.0

# Vector database for contract embeddings
pinecone-client==2.2.4
chromadb==0.4.18
faiss-cpu==1.7.4

# Data visualization for blockchain
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
bokeh==3.3.2
dash==2.14.2

# Real-time blockchain monitoring
websockets==12.0
asyncio-mqtt==0.13.0
blockchain-stream==0.1.0
event-listener==0.1.0

# File handling and storage
boto3==1.34.0
minio==7.2.0
python-magic==0.4.27

# Utilities and helpers
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7
rich==13.7.0
tqdm==4.66.1

# Development and code quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Monitoring and logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0
prometheus-client==0.19.0

# CORS and middleware
fastapi-cors==0.0.6
fastapi-limiter==0.1.5

# Scheduling and automation
apscheduler==3.10.4
celery-beat==2.5.0

# Email and notifications
sendgrid==6.10.0
twilio==8.10.0

# Calendar and scheduling
icalendar==5.0.7
python-dateutil==2.8.2
